+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added `InputType` for setting Python representations of GraphQL Input types
 - Added support for passing `Enum` types directly to `make_executable_schema`
 - Added `convert_names_case` option to `make_federated_schema`.
+- Added support for the `@interfaceObject` directive in Apollo Federation.
 
 
 ## 0.18.1 (2023-02-22)